Finished basic upload feature of FFS

+++ private/views.py
@@ -12,11 +12,11 @@ from django.http import HttpRequest, HttpResponse, FileResponse, JsonResponse, \
     HttpResponseServerError
 from django.utils import timezone
 from django.views.decorators.cache import cache_control
-from django.views.decorators.http import require_http_methods, condition
+from django.views.decorators.http import require_http_methods, condition, require_safe
 from django.views.decorators.vary import vary_on_headers
-from magic.compat import detect_from_filename
 
-from general import default_render, exception_to_response, UserError
+from general import default_render, exception_to_response, UserError, get_mime_type
+from guenthner_xyz import settings
 from private.icons import img_file_icon
 from private.models import FilePacket
 from template_tags import base64e
@@ -40,6 +40,8 @@ def view_index(request: HttpRequest):
 def require_path_exists(func):
     def wrapper(request: HttpRequest, path: Path):
         if not (fs_root / path).exists():
+            if request.method == "HEAD":
+                return HttpResponse(status=404, content_type="text/plain; charset=utf-8")
             return HttpResponse(f"File {path} does not exist", status=404,
                                 content_type="text/plain; charset=utf-8")
         return func(request, path)
@@ -89,9 +91,15 @@ def api_raw(request: HttpRequest, path: Path):
     full_path = fs_root / path
 
     if full_path.is_file():
-        return FileResponse(open(full_path, "rb"))
+        return FileResponse(open(full_path, "rb"), content_type=get_mime_type(full_path))
     else:
-        return JsonResponse({p.name: str(p.relative_to(fs_root)) for p in sorted(full_path.iterdir())})
+        files = list(full_path.iterdir())
+        files.sort()
+        files = [*filter(lambda f: not f.is_file(), files),
+                 *filter(lambda f: f.is_file(), files)]
+        files = {p.name: str(p.relative_to(fs_root)) for p in files}
+
+        return JsonResponse(files)
 
 
 @require_http_methods(["GET"])
@@ -107,11 +115,13 @@ def api_files(request: HttpRequest, path: Path):
         return api_raw(request, path)
     else:
         template_name = "files"
-        title_msg = "Files"
+        title_msg = "Directory"
         files = list(full_path.iterdir())
         files.sort()
-        extra_context = {"files": [f.relative_to(fs_root) for f in files if f.is_file()],
-                         "folders": [f.relative_to(fs_root) for f in files if not f.is_file()]}
+        extra_context = {
+            # "files": [f.relative_to(fs_root) for f in files if f.is_file()],
+            # "folders": [f.relative_to(fs_root) for f in files if not f.is_file()],
+            "net_block_size": settings.FFS_NET_BLOCK_SIZE}
 
     return default_render(request, f"private/{template_name}.html", {
         "title": f"{path}: {title_msg}", **extra_context,
@@ -129,17 +139,16 @@ def api_icon(request: HttpRequest, path: Path):
         return HttpResponse(status=500)
 
 
-@require_http_methods(["GET"])
+@require_safe
 @require_path_exists
 @condition(etag_func=get_path_etag, last_modified_func=get_path_last_mod)
 def api_info(request: HttpRequest, path: Path):
+    if request.method == "HEAD":
+        return HttpResponse(status=200)
+
     full_path = fs_root / path
 
-    try:
-        mime = detect_from_filename(full_path).mime_type
-    except Exception:
-        # Sometimes the function just fails
-        mime = "text/plain;charset=utf-8"
+    mime = get_mime_type(full_path)
 
     return JsonResponse({
         "path": str(path),
@@ -226,6 +235,10 @@ class api_file_ledger(api_class):
         return JsonResponse(status=200, data={"hashes": hashes})
 
     @classmethod
+    def post_status_response(cls, packet_info, success):
+        return JsonResponse(status=200 if success else 202, data=packet_info)
+
+    @classmethod
     def post(cls, request: HttpRequest, path: Path):
         full_path = fs_root / path
 
@@ -233,15 +246,16 @@ class api_file_ledger(api_class):
             return HttpResponse(status=400, content_type="text/plain; charset=utf-8",
                                 content=f"The file at {path} already exists")
 
+        # Client sent wrong body:
         try:
             data = json.loads(request.body.decode())
             hashes = data["hashes"]
         except json.JSONDecodeError:
             return HttpResponse(status=400, content_type="text/plain; charset=utf-8",
                                 content="The request was not valid JSON")
-        except KeyError:
+        except (KeyError, TypeError):
             return HttpResponse(status=400, content_type="text/plain; charset=utf-8",
-                                content="The request did not contain the hashes key")
+                                content="The request did not contain the hashes key (or something related)")
 
         full_path.parent.mkdir(parents=True, exist_ok=True)
         if len(hashes) == 0:
@@ -249,14 +263,16 @@ class api_file_ledger(api_class):
             return HttpResponse(status=200)
 
         missing = []
+        packet_info = {}
 
         for hsh in hashes:
-            if FilePacket.status_for(hsh) != FilePacket.STORED:
+            status = FilePacket.status_for(hsh)
+            if status != FilePacket.STORED:
                 missing.append(hsh)
+            packet_info[hsh] = status
 
         if len(missing) > 0:
-            return HttpResponse(status=400, content_type="text/plain; charset=utf-8",
-                                content=f"The files for {len(missing)} hashes are missing")
+            return cls.post_status_response(packet_info, False)
 
         packets = [FilePacket.objects.get(hsh=hsh) for hsh in hashes]
         files = [p.file for p in packets]
@@ -265,10 +281,10 @@ class api_file_ledger(api_class):
         result = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
         if result.returncode != 0:
-            log.error("Concatenating files on upload failed", result.stderr.decode())
+            log.error("Concatenating files after upload failed", result.stderr.decode())
             return HttpResponse(status=500)
 
-        return HttpResponse(status=200)
+        return cls.post_status_response(packet_info, True)
 
     @staticmethod
     @require_http_methods(["GET", "POST", "HEAD"])
@@ -328,7 +344,10 @@ class api_file_packet(api_class):
         for p in to_delete:
             if p.status == FilePacket.STORED:
                 file = file_packet_cache / p.file
-                file.unlink()
+                if file.exists():
+                    file.unlink()
+                else:
+                    log.error(f"File packet for hash {p.hsh} is missing during automatic cleanup")
             p.delete()
 
     @classmethod
@@ -391,7 +410,7 @@ class api_file_packet(api_class):
 
 @login_required
 @permission_required("private.ffs")
-@cache_control(max_age=settings.CACHE_MIDDLEWARE_SECONDS)
+@cache_control(max_age=4)
 @exception_to_response(UserError, 400)
 def view_api(request: HttpRequest, api: str, path: Path = Path("")):
     valid_apis = ["raw", "files", "info", "icon", "file-packet", "file-ledger"]
